Migrate to using AGP's new variant API in the 'variant' package.
This change adds compatibility classes to support registering an ASM transform in AGP 4.2, 7.0 and 7.1. For 4.2 reflection has to be used for all invocations of the newer variant API since their package in 4.2 is different and its not possible to reference the classes when compiling against a newer version. For 7.0 an onwards, the API are stable and can be referenced directly.

Fixes https://github.com/google/dagger/issues/2700

RELNOTES=Fix an incompatibility in the Hilt Gradle Plugin with AGP 7.0-beta04 and 7.1-alpha02, be aware that with this fix the HiltGradlePlugin will not work with other older alpha/beta versions of AGP 7.0 and 7.1 other than the ones mentioned in these notes.
PiperOrigin-RevId: 380743576